bulk refactoring

+++ app/Produk.php
@@ -13,6 +13,7 @@ class Produk extends BaseModel
 
     protected $fillable = [
             'produk',
+            'slug',
             'kategori_id',
             'kategori',
             'harga',
@@ -33,6 +34,11 @@ class Produk extends BaseModel
     {
         parent::boot();
 
+        static::creating(function ($model)
+        {
+            $model->process();
+        });
+
         static::saving(function ($model)
         {
             $model->process();
@@ -41,17 +47,17 @@ class Produk extends BaseModel
 
     protected function process()
     {
-        if ($this->kategori) $this->kategori_id = $this->kategori;
-        if ($this->brand) $this->brand_id = $this->brand;
-        unset($this->kategori);
-        unset($this->brand);
+        if (request()->has('kategori')) $this->attributes['kategori_id'] = request()->get('kategori');
+        if (request()->has('brand')) $this->attributes['brand_id'] = request()->get('brand');
+        unset($this->attributes['kategori']);
+        unset($this->attributes['brand']);
 
-        if (request()->hasFile('foto'))
+        if (request()->hasFile('foto') && request()->file('foto')->isValid())
         {
             if ($this->foto && $this->id && is_file($exist_file = public_path(static::FOTO_PATH).$this->find($this->id)->foto)) unlink($exist_file);
 
             $destinationPath = public_path(static::FOTO_PATH);
-            $fileName = str_slug($this->produk.' '.date('YmdHis')). request()->file('foto')->getClientOriginalExtension();
+            $fileName = str_slug($this->produk.' '.date('YmdHis')) . '.' . request()->file('foto')->getClientOriginalExtension();
             $result = request()->file('foto')->move($destinationPath, $fileName);
             if ($result) {
                 $this->foto = $fileName;
@@ -98,8 +104,10 @@ class Produk extends BaseModel
 
     public function rules()
     {
+        if (!request()->has('slug')) request()->merge(['slug' => str_slug(request()->get('produk'))]);
+
     	return [
-    		'brand' => 'required|unique:brands,brand'.(($this->id != null) ? ','.$this->id : ''),
+    		'slug' => 'required|unique:produks,slug'.(($this->id != null) ? ','.$this->id : ''),
     	];
     }
